requirements.txt
Django>=4.2.3
neomodel>=5.0.1
djangorestframework>=3.14.0